+++ Log.hpp
@@ -8,15 +8,15 @@
 class Log
 {
 protected:
-    int totalInstructions = 0;                  // 命令の総実行回数
+    uint64_t totalInstructions = 0;                  // 命令の総実行回数
     std::unordered_map<int, int> nStallCount{}; // ストールの回数
-    int branchPredCount = 0;                    // 分岐予測の回数
-    int flushCount = 0;                         // パイプラインのフラッシュ回数
+    uint64_t branchPredCount = 0;                    // 分岐予測の回数
+    uint64_t flushCount = 0;                         // パイプラインのフラッシュ回数
 
     // 命令ごとの実行回数
     std::unordered_map<std::string, uint64_t> instructionCounts;
 
-    int sumCLKCount();
+    uint64_t sumCLKCount();
 
 public:
     void logInstruction(const std::string &instructionName);
